:truck: move board definition
